Refactor(api): Split the three API clients
jenkins_client.py (960 lines), gerrit_client.py (495) and
github_client.py (458) all sat past the 400-line reviewability limit
the aislop code-quality gate applies, and jenkins_client.py also
failed ruff format, which the gate reports separately.

Split each into a package at the same dotted path. The Jenkins client
becomes six mixins its public class composes - discovery, catalog,
builds, matching, jjb and jobs - so method bodies stay untouched and
the MRO resolves every self.<method> call as before. Gerrit splits
into errors, discovery and client; GitHub into status translation,
workflow queries and the client itself. Each package __init__
re-exports the previous names, so api/__init__.py and every call site
are unchanged.

GerritURLBuilder deliberately stays in the package __init__, because
its discover() resolves GerritAPIDiscovery from that module's globals
and two tests patch it there; moving it would have forced a test edit
for no gain.
